+++ src/lang/program_util.hpp
@@ -67,8 +67,6 @@ class ProgramUtil {
 
   static void print(const Operation &op, std::ostream &out, int indent = 0);
 
-  static void exportToDot(const Program &p, std::ostream &out);
-
   static size_t hash(const Program &p);
 
   static size_t hash(const Operation &op);